Remove `before_install` step
We are running all apps on Bundler 2 now so there should be no need to install latest 1.x version of Bundler anymore.

Also, let's stop updating RubyGems. I think it's better that we run the same version on Travis CI as in our DC (and so on).

These changes combined should save us a few seconds.

Close #1